Refactor Icon
